feat: Seprate config page to component
